+++ lizard_history/models.py
@@ -8,9 +8,12 @@ from django.db.models.signals import pre_save as django_pre_save
 from django.db.models.signals import post_save as django_post_save
 from django.db.models.signals import post_delete as django_post_delete
 
+from django.db import models
+
 from django.dispatch import receiver
 from django.contrib.sessions.models import Session
 from django.contrib.admin.models import LogEntry
+from django.utils.translation import ugettext_lazy as _
 
 from lizard_history.handlers import pre_save_handler
 from lizard_history.handlers import post_save_handler
@@ -19,13 +22,17 @@ from lizard_history.handlers import post_delete_handler
 import lizard_history.configchecker
 lizard_history.configchecker  # Pyflakes...
 
-# Django models not to include in history tracking
-EXCLUDED_MODELS = (
-    Session,
-)
 
-# Mongoengin documents not to include in history tracking
-EXCLUDED_DOCUMENTS = ()
+def _is_monitored(sender):
+    """
+    Return if the sender is to be monitored.
+    """
+    
+    test = MonitoredModel.objects.filter(
+        app_label=sender.__module__.split('.')[0],
+        model=sender.__name__.lower()).exists()
+
+    return test
 
 
 @receiver(django_pre_save)
@@ -33,8 +40,8 @@ def django_pre_save_handler(sender, instance, raw, **kwargs):
     """
     Handle django_pre_save signal.
     """
-    # Do nothing when loading fixtures, logging or when excluded
-    if raw or sender == LogEntry or sender in EXCLUDED_MODELS:
+    # Do nothing when loading fixtures, logging or not monitored
+    if raw or sender == LogEntry or not _is_monitored(sender):
         return
 
     pre_save_handler(sender, instance)
@@ -45,8 +52,8 @@ def django_post_save_handler(sender, instance, raw, **kwargs):
     """
     Handle django_post_save signal.
     """
-    # Log nothing when loading fixtures, logging or when excluded
-    if raw or sender == LogEntry or sender in EXCLUDED_MODELS:
+    # Do nothing when loading fixtures, logging or not monitored
+    if raw or sender == LogEntry or not _is_monitored(sender):
         return
 
     post_save_handler(sender, instance)
@@ -57,7 +64,7 @@ def django_post_delete_handler(sender, instance, **kwargs):
     """
     Handle django_post_delete signal.
     """
-    if sender in EXCLUDED_MODELS:
+    if not _is_monitored(sender):
         return
 
     post_delete_handler(sender, instance)
@@ -68,7 +75,7 @@ def mongo_pre_save_handler(sender, document, **kwargs):
     """
     Handle django_pre_save signal.
     """
-    if sender in EXCLUDED_DOCUMENTS:
+    if not _is_monitored(sender):
         return
 
     pre_save_handler(sender, document)
@@ -79,7 +86,7 @@ def mongo_post_save_handler(sender, document, created):
     """
     Log a change or addition of a mongoengine document in the logentry.
     """
-    if sender in EXCLUDED_DOCUMENTS:
+    if not _is_monitored(sender):
         return
 
     post_save_handler(sender, document)
@@ -90,7 +97,23 @@ def mongo_post_delete_handler(sender, document, *args, **kwargs):
     """
     Handle mongo_post_delete signal.
     """
-    if sender in EXCLUDED_DOCUMENTS:
+    if not _is_monitored(sender):
         return
 
     post_delete_handler(sender, document)
+
+
+class MonitoredModel(models.Model):
+    name = models.CharField(max_length=100)
+    app_label = models.CharField(max_length=100)
+    model = models.CharField(_('python model class name'), max_length=100)
+
+    class Meta:
+        verbose_name = _('monitored model')
+        verbose_name_plural = _('monitored models')
+        ordering = ('app_label', 'model',)
+        unique_together = (('app_label', 'model'),)
+
+    def __unicode__(self):
+        return self.name
+